services/social/requirements.txt
notion-client==2.2.1
anthropic
atproto==0.0.61
requests==2.31.0
python-dateutil==2.8.2
python-dotenv==1.0.0
flask==2.3.3
openai>=1.0.0
